Significant folder reorganization and interface update
Far too many scripts in scriptmanager for anyone to reasonably keep
track of. The interface has been modified to display a small text box
next to each script briefly describing its purpose. This was previously
in tooltips embedded in each button, but was generally unknown by
users. The tab organization and naming was also re-done to try and
better fit different tasks. This update is generally pretty
unattractive and will need to be seriously revisited in the future.