Coordinate Manipulation CLI tools working
I filled in the call() and validateInput() methods of the CLI classes
calling these tools. There are more validations to add regarding the
parsing of the user-supplied output paths. For now the tools execute.
Consider merging corresponding BED or GFF versions into the same class
for the scripts/* tools to reduce redundant code.

Missing semicolon added to ScriptManger's main method.