coord-man expand & convert tools CLI work with GUI
Coordinate_Manipulation Expand tools for both BED and GFF files are
working in CLI version, and working for headless mode. Found a way to
print output to STDOUT if no file specified. Possibly need to add more
input checks in the future.

Update build.gradle file so that picocli java file is not needed and
some keywords are changed to be compliant with current versions of
Gradle.