clean up CLI validation in sort coord files
Sort BED/GFF tools had wordy validation code for the CLI side so I
cleaned it up a bit.

Also adjusted parseCDTFile() in window_interface files to be closer to
the methods in the scripts files. They now differ only by the line
notifying the row index of invalid CDT entries. Possibly change output
to String in the future and handle outside the method. Perhaps a cleaner
way exists?