SEStats-separate script from Swing output wrapper
Split the scripts/../SEStats.java file into one of the same name and one
in the window_interface classes as SEStatOutput. In order to do this,
I used the util/CustomOutputStream to create a PrintStream object to the
output swing window in SEStatOutput for the GUI while allowing the CLI
tool to call the SEStats without instantiating any Swing objects.

CLI and Window classes updated accordingly. CLI and GUI both work now
even when jar is run on ACI.