separate out script-PEStat for CLI implementation
Similar to SEStat CLI update, I separated out the script file into a
JFrame SEStatOutput object in window-interface and isolated the
swing-unrelated calculations in the script-SEStat class as static
methods to be called.

Note that the CLI method separates out insert statistics into two
separate text files (insert histogram values and insert
chromsome/summary). Both are displayed together in the GUI and if output
statistics is checked, only the values are output into a file. The GUI
offers no built in way to save the duplicate values or image. All files
are can be saved in the CLI version (Insert PNG, Insert Values, Insert
Summary, Duplicate PNG, Duplicate Summary). The first two are default
and the third can be output with the "-s" tag and the last two can be
output with the "-d" tag.

Furthermore, in order to be able to save the duplicate PNG, the
LineChart class was modified to copy the used method into a version that
include a save as PNG command.