BAMGenomeCorr of BAMstats CLI implementation
Similar to SEStats and PEStats, the script file was  a JFrame object
that was separated out into its JFrame components as a new class under
window_interface with "Output" appended to the end of the name, and its
calculations compenents under scripts which is called directly  by the
CLI class for this tool.

BAMGenomeCorrelation was a little trickier to implement because of the
progress update function using "fireProperty.." that tracks progress in
the GUI needed to be left in the scripts. I commented it out for now so
the setProgress function is not working properly and will need to be
fixed later but both GUI and CLI work and progress can be tracked by
System.out print statements.

As with PEStat, there was an adjustment to the custom charts classes,
specifically charts.HeatMap which involved changing the signature to
require a File input that will save the chart for non-null values.
(analogous to charts used in PEStats)

Also, I forgot to add the PEStatOutput in the last commit so it  is
added
here.