move CustomOutputStream to objects
CustomOutputStream moved from util to objects src directories.

This class wraps a JTextArea object (used by GUI) into a Printstream
object so it can be written to from the script classes. This allows us
to run the CLI version from the same script code without encountering
any swing-related headless mode exceptions.

Of course this involves updating the classes that import this object
so far (SE-Stat and PE-Stat in BAM_Statistics).