fill out CLI class for Fig_Gen/4ColorPlot
Since the 4Color plot was based on a public static void method, this
implementation didn't need as much input for peeling apart JFrame
elements from the script. With this, all Figure_Generation tools have
been updated with a CLI.

However the validation and processing of Color objects took some
troubleshooting. I also updated the HeatMap to accept lowercase letters
in the hex string.

Tested and works on ACI.