cli done for BAMtoscIDX
JFrame script was separated out into a non-swing components of script
and the window components in BAMtoscIDXOutput. The writing of output was
implemented a little differently this time in that I redirected the
STDERR to the PrintStream object wrapping the output window textArea and
used System.err print statements within the script. I similarly  set the
STDOUT to a filename if one was provided.

CLI parsing and validation needed to be updated accordingly and
adjustments were made.

I have not yet checked this tool on ACI but I will check all
BAM_Format_Converter tools together once I've finished with the rest.