cli for all BAM_Format_Converter tools
These script tools were JFrame objects so again I made new
window_interface/BAM_Format_Converter/*Output classes for the JFrame
components and kept the calculations in the script/ classes. I also
altered the BAMtoscidx from the previous commit to print output to
STDOUT back to the way I was doing it for the other classes. Redirecting
the System.out to a PrintStream seemed a little too permanant.

All cli operate similarly. The -s flag indicates that the output should
be directed to standard out. BAMto BED, GFF, and scIDX have this flag
but bedgraph does not because it can output two files: a forward and a
reverse file. It'll get messy if we try to do something with STDOUT. The
filename specified in the CLI is actually just the basename and a
"_forward.bedgraph" and "_reverse.bedgraph" or "_midpoint.bedgraph" will
be appended to it appropriately.

The -f flag indicates "fragment" and is only  available for the BAMtoBED
and BAMtoGFF tools.

Some validation added to the CLI classes for insert size inputs,
conflict between count types, and conflict between the -s and -o
options.

Options in the help are sorted as written in the class, rather than
alphabetically. May do this for future CLI classes.

These tools were briefly tested on ACI.