create CLI for TagPileup
Like others before, this adjustment involved separating JFrame
components from the calculations made. As one of the more complex tools
in this program, the parameter parsing is a bit more complex and
figuring out how to set/validate defaults took a little more time.

Affected classes includes PileupParameters which stores the parameters
used by the TagPileup tool. TagPileupOutput was created as a GUI
intermediate for the results window and TagPileupWindow was adjusted to
call it instead of the script while the script was adjusted to take more
specific filenames to output to (as opposed to generating its own).

The util/ExtensionFileFilter class had a method added to strip a file of
its extension and return the String of the resulting new name.