new CLI tool to create composite plot
The generation of the PNG composite plot figure in TagPileup was
isolated into its own tool under "Figure Generation." Input data series
formatted like TagPileup composite output (first row is x-values and
subsequent rows represent different sets of y-values to plot out like
layered line charts).

The tool automatically detects how many dataseries were input from
parsing the file to determine default color assignments. (Black for one
series, blue and red for two series, and a hardocoded set of colors for
anything more). This tool was generalized to allow for the layering of
any number of lines as long as the data was formatted consistently with
TagPileup output.

Minor edits to main/ScriptManager class to include the new command in
the usage calls and documentation.

The CompositePlot chart class was updated to allow public access of
helper method createChart. This led to the generalization of that class
and color assignments (created a requirement that dataset arrays were
parallel with color array, meaning their number of elements need to
match).

Another edit on the CompositePlot class was on the createCompositePlot
with no color input in the signature method. This was more for
readability and cleaning up redundant code. Functionally there should be
no difference.