cli for seq-analysis/randomize-fasta
Since this tool was implemented with a public static method in the
script class, there was no need for the usual separation of JFrame
compenents. CLI options also minimal. This is a simple tool with
relatively easy implementation.

However I renamed the script for consistency with the naming of the cli
and window_interface classes. (FASTARandomize>RandomizeFASTA)