cli specify filename & validate input/output
Both PeakAnalysis/RandomCOordinate and TileGenome tools are based on
script classes that do not extend JFrame and therefore do not require
the intermediate window_interface/*/*Output class. Although these tools
were already implemented as CLI before, this commit is adding validation
of input and output filenames as well as adjusting the class to write to
specific filenames given in the argument list (rather than allowing the
script come up with a filename based on inputs and parameters).

Affected classes include the window and CLI classes that call the script
method where the default output name needs to be defined and script
object implementation arguments need to be adjusted.