validate input/output and set default filename
The figure generation tools didnt have all the validation of input and
output filenames so I added them with this commit. Also the Heatmap plot
filename to write PNG to is now specified outside the script class so
adjustments to the script and window/*/*Output class were needed to call
the method properly with proper default filenames defined.