fix generalized method for multiple file selection
The method for selecting files was generalized to start adressing the
streamlining of file selection for issue #38. This commit fixes the
method for multiple file selection so that it returns an array of files
rather than just a single file.

One of the file selectors in the ScalingFactorWindow class mistakenly
calls the multiple file selector where I think the single file selector
was intended. This commit fixes that call.