write composites to different files for DNAShape
Due to the different units of the measured values for each aspect of DNA
shape measured, it makes sense to spearate the data into their own files
for ease of composite plot generation. These y-values are not comparable
so plotting them on the same graph is not great for intuitive
interpretation.