create CLI for remaining tools in ReadAnalysis
The two JFrame wrapped tools, AggregateData and ScalingFactor were
adjusted similar to previous commits. A tool "Output" window was created
in the window_interface package to contain all the JFrame components
while the script class was stripped of swing objects that would
interfere with systems not hooked up to a display. CLI parsing for
Aggregate data was a bit more complicated than other tools because of
the merged or individual aggregations.

The ScaleMatrixCLI was updated with the addition of the validateInput()
class for consistency across ScriptManager tools. The script class was
not JFrame-based so the adjustments were less dramatic.

The SimilarityMatrixCLI class is still empty but updated with the
outline that reflects the newer cli class format. This tool requires a
more involved rewrite before incorporation with the rest of the SM
tools.