fix TagPileup tab-formatted output bug
This commit fixes the bug described in issue #41. The problem was in
the case of setting the output to be tab formatted, the Writer
objects OUT_S1 and/or OUT_S2 failed to close.

I adjusted the "if" statements for outputting JTV files (whose code
fell under the same conditionals as the code responsible for closing
the output Writer objects) so that they are consistent with previous
conditions but included the case of PARAM.getOutputType()==1 for
closing the Writer objects.