create CustomOutputStream wrapper object
This wrapper class wraps the GUI's TextArea output object such that the
GUI components are hidden from the CLI version to avoid exceptions being
thrown when there is no way of displaying GUI components.

Since some of the GUI tools output to a TextArea GUI object, running a
CLI version on the terminal of a remote server without adjusting the
script object would throw an exception. By wrapping the TextArea in this
object, CLI won't know that it contains a GUI object unless this object
is instantiated. By writing code to avoid instantiating the object if
running as CLI, we've created a workaround for the GUI text components.

The class extends OutputStream and takes the TextArea object and saves
it in its constructor. The write() method is overriden with the code
that uses the TextArea append() method.