create TagPileup command line version
Before, this tool was implemented with a JFrame script object so this
commit separates out the JFrame elements for CLI implementation. JFrame
elements are saved to the src/window_interface/*/TagPileupOutput class
while the calculations are kept in the src/scripts/*/TagPileup class.

This monster of a tool (at least compared to the others) has extra files
to update and they are described following the first few files that
relate to the general edit structure for tools so far.

window_interface/*Window
	-adjust import to call *Output class
window_interface/*Output
	-adjust import to call script class
	-adjust output argument when initializing script class to be
		full filename and not just directory
script/*
	-strip out JFrame objects
	-use output as filename rather than directory name
cli/*
	-call appropriate script class
	-write up picocli parsing objects
	-write up validateInput() method to check inputs and write
		messages for the user
objects/PileupParameters
	-reorganized methods to put get next to set methods
	-added comments to annotate vars with details about what each
		variable is for
	-renamed a couple methods to be more descriptive by appending
		'get' to the front of the method name
	-added a method that would spit out all parameters as a command
		line execution String. Although unused for now, it may
		be useful for future features.