create CompositePlot command line tool
This tool is unlike the others in that there is no GUI version. It's a
straightforward tool to plot the composite output of TagPileup that uses
the existing chart/CompositePlot class as the script object. Because the
GUI version outputs a composite plot, I created this to retain that
functionality so that when this tool is used with the CLI TagPileup, it
can do the same things as the GUI version.

cli/Figure_Generation/
	-call appropriate script class
	-write up picocli parsing objects
	-write up validateInput() method to check inputs and write
		messages for the user
charts/CompositePlot
	-createCompositePlot() method updated to reduce redundancy
	-createChart(3params) was made public so that this tool could
		take advantage of it, instead of just being an internal
		helper method
	-createChart(4params) was updated to optionally include a legend
		(CLI makes it an optional user-specified value)