create SortGFF command line version
This tool did not implement JFrame elements in the script class so the
CLI implementation did not need to create an src/window_interface/*/
SortGFFOutput class.

There is a quirk with the BEDtoGFF tool that adds a semicolon to the end
of the last column with the feature ID. To make this tool compatible
with the GFF output, I split the last GFF column on the semicolon and
took the first token as the ID.

I also moved the parseCDT method originally in the window_interface/
Coordinate_Manipulation/*/Sort* classes to the util/CDTUtilities class.
The parseCDT method is used to get the number of items in a row of the
input CDT file and to check the formatting for consistent row sizes.
Since both the SortBED and SortGFF tools use the same method, this
commit merges them into the same copy that can be referenced from a
general utilities class. This will make updates easier if this method
needs to be adjusted in the future.

util/CDTUtilities
	-move parseCDT method from SortGFFWindow class to the
		CDTUtilities class
	-add properties to the class to save information on consistent
		row sizes
	-add methods to retrieve these values
	NOTE: static methods of this class appear to be unused or copied
		into another class, consider retiring
window_interface/*Window
	-adjust output argument when initializing script class to be
		full file basename and not just directory
	-lots of differences in this commit are a matter of entabbing
	-rename BED related variable names to be consistent with GFF
		variable naming
	-update CDT file loading to use CDTUtilities validation method
	-remove parseCDT() method since using CDTUtilities version
script/*
	-use output as file basename rather than directory name
	-temporary fix using String's split() method on the semicolon
cli/*
	-call appropriate script class
	-write up picocli parsing objects
	-write up validateInput() method to check inputs and write
		messages for the user
	-import CDTUtilities and parse CDT input to check consistent
		rows as a part of validateInput()