initialize docusaurus for documentation
Using the docusaurus template, this commit structures the documentation
for the tools in ScriptManager. Some preliminary documentation has been
included for each tool and some general documention including a tool
index, some file format specifications, an example analysis of a ChIPexo
dataset, and some basic command line directions have been filled in.

Many tools are still in need of more detailed documentation for the
methodology and many pages are missing descriptions for the tools and
various inputs/options.