minor fix Java warnings
scrripts/PEStats inititalized a Vector without a class for the contents.
This was set to ChartPanel in this commit.

window_interface/FilterPIPSeqOutput stored a boolean value without doing
anything with it upon indexing a FASTA so this commit adds a message for
the user to know if the FASTA was properly formatted or not. This may
need to throw an exception in future updates.