rename heatmap tool
This commit changes the filenames and contents accordingly for the
Heatmap tool in preparation for introducing a new heatmap tool with
positive and negative values in issue #48.

The old heatmap tool will be called TwoColorHeatMap while the new
heatmap tool will be called ThreeColorHeatMap. For the subcommands for
TwoColorHeatMap, we will maintain "heatmap" for backward compatibility
with v0.13 in command line tools.