add new 3color heatmap to gui tools #48
New tool to add to the Figure Generation tool group. A three color
heatmap tool for visualizing matrix/CDT files with positive and negative
values. More details in issue #48.

The main ScriptManager GUI tool lists the three color heatmap tool under
the Figure Generation group just under the original heatmap tool
(renamed TwoColorHeatMap).

The tool descriptions are updated to include one for this new tool.

New tool window and output window classes are modeled after the original
heatmap tool. Users choose three colors (Max, middle, and min) along
with the thresholds for defining each color. Each can be defined either
as a percentile value or an absolute value with error checks in place to
ensure that the middle value is between or equal to the min and max.
Values are also checked to ensure they are within the scope of the
distribution of values in the matrix file. (OptionException)

The script class is largely modeled after the Two color heatmap tool but
instead of scaling between min and max values, the scaling is done
between the min and mid or the mid and max as appropriate.

A new CustomExceptions package was created to house the new
OptionException. This will facilitate more consistent handling of
invalid inputs between the GUI and CLI tools when the CLI version is
created. A similar adjustment in handling invalid options for other
tools may be worth looking into.