add new 3color heatmap to cli tools #48
Add CLI for new tool, ThreeColorHeatMap. More details in issue #48.

The CLI class creates an ArgGroup and class for each threshold (max,
mid, and min) for parsing either a percentile or absolute value out of
the user input line. Colors parsed simlarly to CLI classes of other
Figure Generation tools. Output files named simlarly to the original
TwoColorHeatMap tool with CDT filename appended with compression style.

The CLI class uses the recently created CustomExceptions.OptionException
class.

The script class was updated so that other compression styles throw the
custom exception when max/mid/min are invalid (instead of just printing
to STDERR it was doing before).a

The Output class was just stripped of an unused import.