streamline version updating
Instead of updating the version in multiple locations and instead of
only running the main JAR file (not the subcommands) for the command
line version statement, this commit updates all classes so that version
incrementation needs only to adjust one shared variable in
ToolDescriptions and the version update will propagate to all commands,
subcommands, and the main GUI ScriptManager window title.

The static final variable containing the version string in
ScriptManagerGUI was moved to objects.ToolDescriptions which contains
static strings contained in an object that all command line interfaces
import. It is also imported by the main GUI JFrame so no new import
statements need to be added and just the addition of a `version =` line
in the Picocli @Command statement.