restructure tools using .fai to throw exception
For the three tools that require a FASTA index file, the FASTAUtilities
class is used to check the FASTA format and create an index file. This
commit creates a new custom exception object (FASTAException) to be
thrown by FASTAUtilities when a FASTA file with bad formatting is input.

The tools affected are DNAShapefromBED, FASTAExtract, and
FilterforPIPseq.

All script tools are adjusted to check for the presence of an .fai file
in the constructor and generate one if necessary. They will also throw
FASTAException or IOException objects.

Output window classes will pass along the FASTAExtractt objects and
reduncant FAI checks in these objects are removed.

CLI and Window classes will catch the FASTAExceptions and print an error
statement or pop up a dialog box for the user.