adjust var names and fix error message window
Both the DNAShapefromBED and FASTAExtract tools throw message dialog
boxes with "No BAM files loaded" but since both tools are based on BED
inputs, the message is adjusted with this commit.

Window object var names are renamed to be more descriptive by adjusting
from BAM-based names to BED-based names.

Output directory path is switched from OUTPUT_PATH to OUT_DIR to be more
descriptive.

Unused INDEX variable is stripped. This was a remnant of shift from
boolean-returning FASTAUtilities method to an exception-throwing one.