rename output vars and fix SearchMotif default
This commit starts a shift to a more descriptive naming standard that
differentiates File objects for the directory used in the window classes
from the full filepath with the filename for the output.

Output filenames are defined within the Output class.

SearchMotif tool was fixed so that the GUI version of the tool writes
to a proper output file.