rename output vars for read-an tools
This commit renames output variable names to be more descriptive within
the window classes. The goal is to differentiate File objects for the
directory from the ones used for the actual filename of the output.

This will also add consistency across tools in the way the output
filepaths are built.

The PileupParameters methods were renamed to be more descriptive in
distinguishing output directory file objects from filename objects.