bugfix tagpileup was not outputting composite file
Several exceptions were being thrown and the composite plot output file
was not being written. The two primary changes are elimination of unused
COMPOSITE variable in the TagPileupOutput class and the change of the
OUTPUT directory default in the TagPileupWindow class to be the user
directory that ScriptManager was executed from.