add checkbox for chrmt option #49 ConvertChrNames
As described in issue #49, different sources for the sacCer3 use
diffrent chromosome naming for the mitochondrial chromosome. As a
result, we are adding the option to conver the 'chrM' to 'chrmt' with a
checkbox option in the GUI and a option flag in the CLI.

src/cli/File_Utilities//Convert*ChrNamesCLI
-option for `-m` or `--chrmt` added for swapping out "chrmt" --> "chrM"
map to "chrM" --> "chrmt" map.
-script call adjusted to include boolean argument
src/scripts/File_Utilities/ConvertChrNames
-boolean parameter added to A2R and R2A wrapper methods
-helper methods that generate chromosome name HashMaps for converter
tools conditionally add one of the two mitochondrial name maps based on
new input boolean parameter.
-workhorse chr converter method unchanged, only input HashMap adjusted
based on input boolean argument.
src/window_interface/File_Utilities/Convert*ChrNamesWindow
-checkbox objects are added while other window objects are shifted and
adjusted in th window layout to accommodate
-script call added extra input argument