upgrade docusaurus to v2.73 using npm
The docusaurus version has been incremented in the package.json file
with this commit.

The yarn package management file (yarn.lock) is removed and replaced
with the analagous npm file (package-lock.json).

This should address a handful of the dependabot security alerts.

Switching package management (lock files) to npm rather than yarn. This
will match the package manager used in other CEGRcode repos. While there
is some performance loss with npm, it appears to be more secure as well.