separate chart generation from script TagPileup
Only the GUI version of TagPileup generates the composite chart so the
classes for this tool arre restuctured in this commit so that the chart
is generated by the window_interface/TagPileupOutput class. This results
in one less argument input to the script so the CLI class was also
minorly affected by this commit.