output composite via printstream to fix overwrite
Problem: When the gui wrote the composite output to the
composite_average.out file, runs with multiple BED x BAM combinations
would be sequentially overwritten.

This was fixed by changing the PileupParams composite output object to a
PrintStream that could be initialized at the TagPileupWindow object
before iterating through each BAM and BED file. The get and set methods
for the PileupParams were updated accordingly for all classes that used
the get/setCompositeFile --> get/setCompositePrintStream.