deprecate JTV File Format #64
Affected tools are TagPileup in Read_Analysis and the two sort tools in Coordinate_Manipulation

Remove the util.JTVOutput class that writes a JTV file.

Remove the code within the SortBED and SortGFF script classes that call the JTVOutput class to generate the JTV.

Remove the code within the TagPileup script class that calls the JTVOutput class to generate the JTV.

Remove the JTV-related variables stored in the PileupParameters class that TagPileup uses.

Remove the code within the CLI and window_interface class that set the PileupParameter values.