move composite plot color information from PileupParameters object
Since the JTV file type was deprecated, the command line interface of Tag Pileup does not use color information because it only stores the values of the composite plot rather than generating the PNG (see separate composite plot tool).

As a result, the PileupParameters object is used by the script TagPileup object and the colors are used only by the Output window from the GUI interface. It is redundant to pass the color information to the script when it is only used by the output window.

PileupParameters.java--color information (Anti, Sense, and Combined) are removed from the object.

TagPileupOutput--constructor adjusted to accept an ArrayList of Color objects and the composite plot figures are adjusted to use this ArrayList for generating the composite plot object. Two whitespace formatting changes are also included here.

TagPileupWindow--The code for changing the PileupParameters color fields is adjusted to create a new ArrayList of Color objects to pass to the TagPileupOutput object.