remove input file extension restriction
This code has previously been written to restrict the file extensions used for inputs and outputs for enforcement of good file naming practices but in order for it to be compatible with Galaxy which names all files with ".dat" extensions, this commit lifts this restriction. The lines checking the input file extension and outptu file extension of each of the coordinate file converter tools are removed.