fix TP BEDcoord-strand shift
By removing the code block shifting the BED coordinates that appears to be some sort of strand-specific correction, we restore the data to the correct composite (checking for BEDcoord with different strands and checking both sense and antisense composites). Suspect the correction code block that is removed in this commit was for correction of code that is no longer in place.

Relates to issue #75