TP change insert size calculation
The rationale is outlined in issue #76.

PileupExtract is updated for pileups that require proper pairs.

The midpoint is changed to get the leftmost coordinate of the insert (getAlignmentStart or getMateAlignmentStart depending if R1 or R2 is the leftmost read) and add on half of the insert size.

Note the correction for when the BED interval is even and on the negative strand. This ensures that even though we perform a floor integer division calculation, the distance from the 5' end of the BED interval is consistent between BED intervals. Odd intervals have consistent 5' distances regardless of direction.

The filter for insert size is changed to use the built-in SamRecord function getInferredInsertSize() by checking if the absolute value is more or less than the limits specified by the PileupParameters object.

I also switched the filter to use a continue statement instead of setting FivePrime to an invalid position in order to save a little on downstream computation.