rename variables in PileupExtract
Two variables renamed in PileupExtract: BEDCoord `read` renamed to `coord` and int variable specifying marked location of for pileup renamed from `FivePrime` to `mark`

The BEDCoord is renamed to specify it as different from the SAMRecord objects.

The int type variable is renamed `mark` because the variable is/will be used to mark features other than just the FivePrime end including the three prime end and midpoint.