Create gradient-colored buttons for picard MarkDup
This commit includes a new GradientButton object which acts like a regular button except when you color the background it will be colored with a gradient from transparent to the background color. JButton default background coloring creates a flat color which is inconsistent with the metal look-and-feel and can potentially create confusion over whether or not a button is "pressed."

MarkDuplicates is the first tool that directly uses the Picard implementation of one of its command line tools so we will color it orange in this commit to distinguish it from other custom build tools that we've written. Going forward, future Picard tools should be marked/highlighted consistently.