minor README updates
The wrapper README file was updated to acknowledge Audrey and Ali for their contributions, mention that the built ScriptManager `.jar` file needs to be added to the `galaxy/wrapper` directory, and describe the current status and goals of the file contents in the `galaxy/wrapper` directory.

The workflows README file was created to mark the workflow directory in Github and describe the workflows that we plan to include.