update expand-bed docs
Write up documentation for the Expand BED tool as one of the prioritized tools used in ChIP-exo analysis. This includes brief descriptions of the different user-specified options and details on the expansion strategy and use of the "Size of expansion" variable. Draft figure is included and will need to be redone. Will need to merge in with Expand GFF when the tools are consolidated.