generalize docusaurus figure gallery code
The ChIP-exo, Genomic Features, and ATACseq tab of the figure gallery were a mix of hardcoded formatting. This commit creates a new "GalleryTab" function that inputs a single AssayInfo object and builds the gallery tab structure. All three assay can use this structure to display in the gallery. The image urls can be one or more (arranged using flex display).

Future work: consider turning "GalleryTab" into a component (`src/components/GalleryTab`) and fixing tutorial link to dynamically check for an href value to the tutorial.