add optional "includeGZ" feature to FileSelector
Relating to issue #91. Add code that supports properly checking file extensions such that the file selector can be toggled to include the gzipped versions of the file extensions.

Essentially the version of the FileSelector that includes gzipped files will check for a ".gz" extension and then match the extension preceding the ".gz" to the accepted extension options. This optional File Selector ensures other tools leveraging the default version will not include gzipped files in the list of options before we can implement code that supports GZipped files.