add list of publications using SM
Ying created a JSON list of publications that used ScriptManager in their analysis. This commit formats the JSON into a table in a markdown file on the website docs.

The Tool Index table is also moved to the same "References" directory to reflect the sidebar organization of the website.

Links between the "Citing us" section and the list of publications that used scriptmanager are added so they point to each other