split CompositePlotCLI to create script class
For the purpose of building out a GUI for the composite plot image generating tool (#96), the CLI class is broken up in this commit to separate out the input data parsing and chart building from command line parsing and input validation components in the CLI class. The new class organization further added minor adjustments to script behavior:
- the default filepath suffix is now "_plot.png"
- color defaults are determined by a null input and n>2 condition uses YEP color palette instead of Kelly colors
- data parsing issues throw Exceptions rather than print statements
- input/output extensions no longer restricted

The new ColorSeries class saves color palettes and can be reused and expanded upon for future tools.