Added the DownsampleSam Picard Tool
Matched this branch with changes made to the dev branch when Erik-Branch was merged. This will allow a simple merge in the future as this branch is basically a fork of dev now.

Added new Picard tool DownsampleSam. The tool is in the DownsampleSamFile class and its supporting window class is called DownsampleSamFileWindow

Added scroll functionality to the BAM Manipulation window in allow the user to select DownsampleSam