Added SamFormatConverter and SamtoFastq
#122 Added two new tools. SamFormatConverter converts a SAM file to a BAM file or vice versa. SamtoFastq converts a BAM file to a FASTQ file. It has two features both of which are experiencing errors causing no output. However the default configuration runs fine.

Did some minor code cleanup on ValidateSam as well. Along with an additial condition being handled in DownsampleSam. Before, when a user entered a desired maximum number of reads to be outputted that is larger than the total number of reads, the program would default to 0.5 probability without alerting the user. A JOptionPane now tells the user this fact.

Also, added the option to load a SAM file instead of a BAM file for DownsampleSam and ValidateSam.