Added New Tools
Added NormalizeFasta and CollectBaseDistributionByCycle Picard tools and added the wrapper for the FilterSamReads tool. Refactored BAIIndexer to BAIIndexerWrapper and Window to match with other tools.

The charting feature for CollectBaseDistributionByCycle experiences an unknown error on my end. I am unsure if this has something to do with my R install or the code itself.